init structure for connector api
