+++ src/api/connectors/SpotifyConnector.ts
@@ -1,0 +1,1 @@
+// parse spotify payload and convert it to internal representation to store it finally in the DB
